back/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.1
Django==4.2.6
django-environ==0.11.2
djangorestframework==3.14.0
idna==3.4
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
cffi==1.15.1
cryptography==41.0.3
defusedxml==0.7.1
dj-rest-auth==4.0.1
django-allauth==0.54.0
django-cors-headers==4.2.0
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
requests-oauthlib==1.3.1

